# docs/qa/openapi.yaml
...
openapi: 3.0.3
info:
  title: Stitch ATS API
  description: "REST API for Stitch ATS — recruitment, candidate pipeline, interviews, offers, and portals.\n\n**Authentication:** JWT via `POST /api/auth/login`. Pass `Authorization: Bearer <token>` on protected routes.\n\n**Integration key:** Microsoft 365 routes use header `X-Integration-Api-Key`.\n\nSee also: [API_REFERENCE.md](./API_REFERENCE.md) for QA test accounts and workflows."
  version: 1.0.0
  contact:
    name: Stitch ATS
    url: "https://stitch-ats.in"
servers:
  - url: "https://qa.stitch-ats.in"
    description: QA staging (frontend proxy)
  - url: "https://ats-0dtj.onrender.com"
    description: QA API (Render direct)
  - url: "https://stitch-ats.in"
    description: Production (frontend proxy)
  - url: "https://stitch-ats.onrender.com"
    description: Production API (Render direct)
  - url: "http://localhost:3001"
    description: Local dev API
...
paths:
...
  /api/portal/positions:
    get:
      tags:
        - Candidate Portal
      summary: Open positions
      description: "**Roles:** CANDIDATE"
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Position list
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/portal/positions/{id}:
    get:
      tags:
        - Candidate Portal
      summary: Position detail
      description: "**Roles:** CANDIDATE"
      security:
        -
          bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Position
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/portal/me:
    get:
      tags:
        - Candidate Portal
      summary: Candidate dashboard
      description: "**Roles:** CANDIDATE"
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Dashboard
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/referral-portal/me:
    get:
      tags:
        - Referral Portal
      summary: Referrer dashboard
      description: "**Roles:** Referral portal (not CANDIDATE or VENDOR)"
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Dashboard with referral code
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/referral-portal/positions:
    get:
      tags:
        - Referral Portal
      summary: Referral positions
      description: "**Roles:** Referral portal (not CANDIDATE or VENDOR)"
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Positions
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/referral-portal/positions/departments:
    get:
      tags:
        - Referral Portal
      summary: Position departments
      description: "**Roles:** Referral portal (not CANDIDATE or VENDOR)"
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Department names
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/referral-portal/positions/{id}:
    get:
      tags:
        - Referral Portal
      summary: Position detail
      description: "**Roles:** Referral portal (not CANDIDATE or VENDOR)"
      security:
        -
          bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Position
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/referral-portal/referrals:
    get:
      tags:
        - Referral Portal
      summary: My referrals
      description: "**Roles:** Referral portal (not CANDIDATE or VENDOR)"
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/referral-portal/referrals/{id}:
    get:
      tags:
        - Referral Portal
      summary: Referral detail
      description: "**Roles:** Referral portal (not CANDIDATE or VENDOR)"
      security:
        -
          bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/referral-portal/parse-resume:
    post:
      tags:
        - Referral Portal
      summary: Parse resume
      description: "**Roles:** Referral portal (not CANDIDATE or VENDOR)"
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ResumeUpload"
      responses:
        200:
          description: Parsed fields
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/referral-portal/check-email:
    get:
      tags:
        - Referral Portal
      summary: Check email
      description: "**Roles:** Referral portal (not CANDIDATE or VENDOR)"
      security:
        -
          bearerAuth: []
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: Existence
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /api/auth/login (~7 day lifetime)
    integrationApiKey:
      type: apiKey
      in: header
      name: X-Integration-Api-Key
      description: M365_INTEGRATION_API_KEY env var
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
        code:
          type: string
      required:
        - error
...
    ResumeUpload:
      type: object
      properties:
        resume:
          type: string
          format: binary
          description: PDF, DOC, or DOCX — max 5 MB
      required:
        - resume
    Candidate:
      type: object
      description: Full candidate record with pipeline fields
      additionalProperties: true
...